Create Blocks and Work Trackers from DashboardV2
When you click anywhere in the main scrollable grid, it will open a modal to edit block text for the selected cell. This modal is called CellEditor.tsx and it has a button to go to the work tracker. Then this button will open the work tracker modal that we've already been using.