Cell Renderer added
So Grid is the generalized class that displays a grid. Not virtualized or baking yet. It takes a CellRenderer Interface as a prop to the constructor, so then when you want to define the rules you will create a class that implements this interface and returns a container with all the objects that you want to add to this cell.