added header grid
this is proving that the grids also work out of the box as sticky rows or columns because they won't scroll if they don't have any content. They are also listening for scroll wheel events to update the position so you can scroll with you mouse over any of the sticky rows or columns.

Also added a param to showScrollbar default to true but can also hide in sticky row case so we don't see the scroll bar multiple times.