Added Virtualization
So we've added virtualization to the grid now and it seems to be working. I've tested FPS differences between a 40 x 40 grid and a 40,000 x 40,000 grid and FPS is the same.

Only problem is it's still slow. I can see that if we scroll around a tiny amount so the cell pools don't update, then it's buttery smooth but i think its redrawing everything every time we scroll far enough to re-render rows / columns.