Pinned Labels (SLOW)
I think I figured out the start to pinned labels. I removed all the baking stuff so it's really slow only rendering entire graphics.

The logic is to setup a new mainGrid over the existing grid with the same dimensions and rows, but only one column. This creates a new layer essentially that shares all the properties of the grid but will only scroll vertically.

So then we can make a new CellRenderer that only shows the Labels for each event because these are the only things that will be pinned and we use the firstVisibleColumn data from the mainGrid to conditionally display the pinned labels.

Now it's time to re-incorperate the baking feature without losing interactibility again.