Minor Performance Improvements
I think the big thing was building one less container per cell. Each cell renderer would create a container no matter what and put a sprite in it then return the container with a sprite in it and the grid would add this to it's cell container. Reducing a single container build here means we can build like 1000 less if we're zoomed out. Also we're only listening to scroll position on one grid now.