HoverableBakedSprites - Here now!
So this worked out really nicely. If we want to have a complex object that will be baked unless it's hovered, then we can use a HoverableBakedSprite class!

Just create a new class that extends HoverableBakedSprite (which extends a Container btw) and define the callback function in the super with a key and optional dimentions and the hoverable baked sprite will take care of reacting to hover events!

I'm pretty sure you could nest these too if you wanted to but idk if you need to.