vertical scroll less content after refresh fix
the issue I'm having, is if the content of the grid's height is less than the screen height, it doesn't allow scrolling but sometimes the grid refreshes with less rows so we're under all the rows but cannot scroll back up to the top.

made it snap to the appropriate y position or the bottom of the content. which ever comes first. obviously if the content is less than the viewPort height, then just set the y scroll back to 0.