loading events properly
This is a bit of a bug fix. we were using the global table stores to get the data to set the current event when clicked. but they weren't updating fast enough so i could update an event and then click on it again and it would look like my changes weren't saved but I just needed to refresh.

So now it'll query the DB to load the event.